Move mixed_integer low level functions in utils
